requirement.txt
fastapi
uvicorn
dotenv
pdfplumber
pyjwt
python-multipart
psycopg2-binary
numpy
sqlalchemy
google-generativeai